Create ortholog_combinations.py
Creates a batch script for SLURM to find orthologous introns between all possible pairs of introns in the core database. If you want to add multiple new genomes to the database, you need to add them one at a time and add their names to the lists once each one is added.